# MAINT: Improve docstring and performance of trimseq