* The docstring of `numpy.polynomial.polyutils.trimseq` stated "This routine fails for empty sequences.". This is incorrect, as `trimseq` works fine on empty sequences (there is in fact an if statement especially for this case). In this PR we modify the docstring and add tests to make sure the `trimseq` works fine on empty sequences.
* The most common case for `trimseq` are non-empty sequences (the coefficients of polynomials are non-empty by construction) with final coefficient non-zero. We add a fast path for the case where the final coefficient is non-zero.

Benchmark
```
from numpy.polynomial import Polynomial
from numpy.polynomial.polyutils import trimseq

p = Polynomial([1,2,3])

%timeit trimseq(p.coef)
``` 
Results:
```
main: 565 ns ± 36 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
PR: 224 ns ± 21.1 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```
<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      https://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      https://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      https://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->