@eendebakpt 2023-08-22
> Seems fine to me - the implementation change is not likely to have performance implications in practice, but AFAICT it doesn't hurt! +1 for aligning the docstring with the behavior.
> 
> Any objections @charris ?
> 
> While we're at it - I took the liberty of correcting a minor defect in `test_trimseq`. I suspect the intention was to test with varying numbers of trailing zeros rather than run the same test 5 times.

Thanks for fixing the test in `test_trimseq`. I agree that performance gains on the polynomial operations itself are small (but they are measurable), together with several other improvements I hope to make some gains.