@charris 2023-08-22
Thanks @eendebakpt . I'm not sure and empty sequence is useful, but at least the documentation is consistent :)